Move sqlite and mysql config into a config namespace
Previously, the mysql config was coupled with
keystone::mysql. This was problemtatic b/c the
config is for a database client which may not
reside on the same server as the db itself.

This commit creates a new config namespace and
puts the db specific config there.

It also moves the sql template to mysql