Make keystone error handling and retry mode robust
When keystone starts, there is actually a delay before
the keystone client can connect to it.

There was code that captured the appropriate error message
for Ubuntu, but this was not propery handled or Redhat.

This commit changes the code so that any 400 error will result
in one retry. It also reduces the sleep time before the retry to
10 seconds.