Correct keystone client connection defaults
This patch resolves a few minor issues related to defaults.

It resolves tests and docs related to a misconfigured port default
that has already been corrected.

It also changes the protocol default from https to https for client
connections. This is b/c the default protocol for keystone is http.