Add keystone code to capture and ignore warning
When upgrading to grizzly, I found many errors
in the keystone command line output.The provider
was unfortunately not robust enough to still
parse STDOUT when these error appeared.

This commit wraps all keystone CLI invocations
in something that filters out warnings.

This patch is intended to resolve the issue in a way
that still maintains backards compatibility with folsom.