Select keystone endpoint url bases on SSL setting
When keystone is ssl enabled, (by setting ssl/enabled = TRUE in
keystone.conf) the keystone auth server is using HTTPS instead of HTTP
as the protocol. This patch generates the correct endpoint for the
keystone service calls.

In addition, if there is a defined 'admin_endpoint' in keystone.conf,
this setting will take precedence.

Change-Id: I23fe552808d8a959b0dff1c86989cf2aaa04582a