Adds Support for Managing the Token Driver Backend
Previously the keystone class did not manage the token driver.
High-Availability deployments of Keystone must use the SQL
driver so that all Keystone instances can access the same token
from a SQL database cluster.

The patch adds the token_driver parameter that is used to manage
the selection of the token driver backend in keystone.conf.
The parameter defaults to 'keystone.token.backends.kvs.Token' for
backwards compatibility and to use the KVS backend driver.

Change-Id: I4879052b7e44272c176814d048f916cb77d33581