Do not set tenant description in cache
Prevously, after the full list of tenants
was retrieved to build the cache, an
extra call was made to tenant-get per tenant to
retrieve the tenant's description.

This extra call changed the cache creation
operation from O(1) to O(N).

This commit improves the query complexity to
0(1) by moving the description lookup from
part of self.instances to the description
getter method (essentailly making it lazy).

This will work different depending on if you
are interacting with the resource in query
vs. manage mode.

In query mode, self.instances will build
out the cache withouth the description initially.
Since description is a properly, it will invoke
the getter method.

In manage mode, the description getter will only be invoked
for cases where the description is managed.

Partial-Bug: 1224179
Change-Id: Id9268678e122ba3aa3e5b0bcfd77466071a4192d