Don't set keystone endpoint by default
In the current version of the keystone::init class, the default
values for keystone endpoint URLs are unversioned, which can lead
to problems connecting to the endpoint.  This patch reverts to
the old behavior of not setting them at all by default, and includes
sample values with versioned URL's both in the header and tests
so that users know to use versioned URL's if they choose to
override the defaults.  It also updates the documentation in the
header accordingly and tidies it up a bit.

Change-Id: Ib14edf152c1d208418f101ed48cdc38b18f840a2
Closes-Bug: #1287520