enabling tokens flushing using cron
This patch is for configuring a flush token cron when using database as
backend for tokens storage.
While potentially useful for auditing in production environments,
the accumulation of expired tokens will considerably
increase database size and may decrease service performance,
particularly in test environments with limited resources.
It's recommanded to create a crontab to flush tokens every hour.

Change-Id: Icd819cc95b3440f43c5b588f9f2dd6ca47b25a71
Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>