Use UUID as the default token provider
Keystone changed it's default token provider to UUID for Juno.
This patch changes the puppet-keystone default to match.

One of the existing spec tests was incorrect, and had to be removed.
The test was expecting signing certificate setup to be performed wen
enable_ssl is True, but that setup is really only performed when
configuring PKI format tokens.  There is already a test that covers
properly enabling SSL and PKI token format configuration, so this
incorrect test can be removed.

Change-Id: Id838bbb06602b9fe5331329c74872b62932ace1b
Closes-bug: #1372991