handle missing project/tenant when using ldap backend
The project/tenant is only stored with the user when using
the sql (default) backend.  This means calls like user list
and user show will not show the default project.  For getting
the project/tenant, if it is not present in the instances
information provided by user list --long, we don't actually
get the project/tenant, we just check to see if the user is
assigned to the given tenant/project.  For setting the
project/tenant, this is usually done with user create, but when
ldap is read-only, we have to use role add to add the user to a
default _member_ role and at the same time assign the user to a
given project/tenant.  This also adds support for using the
openstack $obj show commands which return --format shell output,
and return a hash of key/value pairs.

Closes-Bug: #1391373
Change-Id: I9a4f72f9a1de200b324aa52eb3ea90d15423c062