Add max_token_size optional parameter
This adds a param to set the max_token_size in keystone.conf.

The setting allows larger tokens (i.e. PKI and large catalog) to work,
in line with other projects that have increased the default from 8k to
16k max. The default setting in Keystone is 8k, the default in this
change is undef, which removes the config item from keystone.conf.

Change-Id: I1e73cf274c26d1368973207a5823ca7a6197612b