Undeprecate keystone::endpoint class
This class should not be deprecated in favor of
keystone::resource::service_identity as we have no way
to configure keystone::resource::service_identity through Hiera yet.

It would be the equivalent of deprecating all *::keystone::auth
in all our modules which was not our intention by introducing
the common keystone::resource::service_identity resource.

Change-Id: Ib424ac9b2c1e61e1e39fe56fd1adfe552110fcb8