Test keystone_user password with Net::HTTP
Openstackclient 0.3.0 does not have the `token issue` command. If the
user is not an admin then there are no other keystone-related actions
the user can perform with the openstack command in order to verify
their password. This patch changes the password verification step to
make an HTTP request directly against the API. Since there is no
session state that needs to be managed, and because this is a special
type of request, we do not need to go through the
Puppet::Provider::Openstack request methods in order to accomplish this
verification.

Closes-bug: 1408754

Change-Id: I45f364efb585e09412de4398cc3f65615ebc5bb7