Add timeout to API requests
Sometimes a service is temporarily unavailable, for instance
immediately after puppet refreshes it. This causes the providers to
fail when it would be optimal to wait briefly for the service to
return. This patch adds a timeout loop in the style of [1] so that the
keystone providers can continue to attempt requests if the service was
unavailable during their first attempt.

[1] https://review.openstack.org/#/c/130155

Change-Id: I33ea824a47ca5834b3f97aa9a714ec385984f00f
(cherry picked from commit c6f70aeab01651e2b821acb75c5b4c253df3bf6a)