Set WSGI process display-name
The display-name option for mod_wsgi's WSGIDaemonProcess directive
allows for a useful name to be used for the process command name.
This is generally useful, as it allows one to distinguish the
keystone wsgi processes from other httpd processes on the system in
the output of ps and similar commands.

This patch uses display names of 'keystone-main' and 'keystone-admin'
for the main and admin processes respectively.

Change-Id: Id79c3a36b3f84938070f41c7171dfee20d1745e9
Closes-bug: #1409119
(cherry picked from commit 7108a6e64f8e3c2c7586ce74b6af686dc16c6a28)