Shorten HTTP request timeout length
If a service is really not available and not just in the middle of
refreshing, it will be apparent relatively quickly. There is no need to
have the user wait 120 seconds to fail. This patch decreases the
timeout length to 10 seconds.

Change-Id: Ia9e13bd6c6f867b00f0a14ccff86cb3ab52f9fe3
(cherry picked from commit eb4f43e25cc40160980bf1fc45a48e1abe147d65)