Implement caching lookup for keystone_user_role
The new implementation of keystone_user_role failed to implement a
previously existing method, user_role_hash, which helped cache the user
role matrix. This helped prevent subsequent runs of exists? from
re-requesting full lists of projects, users, and roles for every found
instance of keystone_user_role. The failure to implement this caching
causes extremely slow puppet runs even when no changes were needed
since the provider would rebuild the hash for every call of exists?,
which even for a small number of users and projects is very time
consuming.

Additionally, the new implementation introduced a bug in the destroy
method where it might cause a failure if the user had some role in a
project, but not all the roles in the roles property of the resource.

It was also realized that since this provider does not make use of the
flush method, the initializer was not needed.

This patch addresses the issues with the following:

- Adds a class method user_role_hash() for caching the matrix as a class
  instance variable so that the value found in self.instances() can be
  used in the next call to exists?()
- Adds a class method self.set_user_role_hash() for setting the class
  instance variable so that instance methods can set it
- Attempts to get the cached value in self.build_user_role_hash() and
  sets it after making the API request if it was empty
- Changes the exists? method look up the cached value from
  self.instances if it is there, and if not, make only one request for
  the user role list for the given project and use those values to
  create the property hash
- Changes the destroy method use the value from the property hash
  instead of the resource to determine what roles to remove
- Changes the roles method to use the value from the property hash
- Remove the object methods 'instances' and 'build_user_role_hash'
  since there is no longer any reason for an instance method to try to
  build the whole hash
- Change the object method 'instance' to use the cached hash instead of
  attempting to rebuild it
- Removes the initialize method since @property_flush does not need to
  be initialized in this provider.

Change-Id: Ifa67ef60e536f88bded0e8db984728f9bd76b75c