Rename resource instance variable
Currently, the method instance() returns a single instance of the
resource in question for keystone_user, keystone_tenant, etc. It caches
the result in a variable called @instances. This is a confusingly named
variable since it only holds one instance. This patch simply changes
the name of the variable. It should not have any effect on any
functionality as nothing else uses the variable.

The keystone_user_role resource will implement caching differently [1]
so this change is unnecessary for that resource.

[1] https://review.openstack.org/#/c/150200/

Change-Id: I21dc22876f47d99108aaf5e90e514682ee125140