Create resource cache upon creation
Currently, the keystone providers ignore the output of the create
command even though it provides much-needed information. This is a
problem because if during the course of the puppet run that
information is needed again, it must make another API request for it.
This is especially problematic with the neutron providers, which
instead of looking up the information via the API, look it up in the
catalog. Since a keystone resource does not specify its ID upon
creation, it reports it as nil, causing the neutron providers to
silently fail when they need the ID of a keystone resource.

This patch corrects the problem by saving the output from the create
command in the @instance variable so that it is known to that
resource and can be retrieved by other resources.

Alternative to https://review.openstack.org/#/c/154812/

Change-Id: I2090f5f4a94a6ae21bf71ac0f1b33caaa602e0bc
Closes-bug: 1421573