Remove warnings from openstack responses
Puppet's `commands` and `has_command` methods are not able to separate
stdout and stderr. When the openstack command prints out unexpected
warnings that have no bearing on our computation, we need to ignore it,
otherwise the CSV module will fail to parse it. This patch filters the
total response produced by the openstack command into only lines that
match a CSV-like regular expression with at least two fields.

If a legitimate error is produced that causes the command to return
failure, puppet catches it during the command execution and raises a
Puppet::Execution error, so this change will not affect its ability to
detect legitimate failures and report on them.

Change-Id: I4f3a8b1ddf6b06dcbf194dd66408c71d4da6d861
(cherry picked from commit 24effa1e360b1eacec7fbfdfed339ba6e639f7fd)