Fix ipv6 support
There is issue with ipv6 where address has to be in brackets, this causes the
underlying ruby TCPSocket to fail. Net::HTTP.new will fail without brackets on
joining the ipv6 address with :port or passing brackets to TCPSocket. It was
found that if we use Net::HTTP.start with url.hostname the incriminated code
won't be hit.

Change-Id: If9b2b629b701c4a4c2794c3b4af584bcfc47fc2d
Closes-Bug: rhbz#1185652