support for keystone v3 api - the project resource
This patch implements these parts of the blueprint:

1) Implements the domain aware keystone_tenant (project) resource.

Adds support for a :domain property of a tenant/project, and allows
the domain to be specified as "name::domain" in the :name property.
The new code tries to use the id as much as possible, since it is
the id that uniquely identifies the project/tenant among all of
the domains, and using id means that name + domain are not required.

Implements: blueprint api-v3-support

Change-Id: I1f703affa6172a3e759e2d62422a87073262acc4