support for keystone v3 api - keystone and keystone::roles::admin
This patch implements these parts of the blueprint:

1) Allows user to specify the default domain
If the user specifies a domain other than 'Default', class keystone
will create that domain, and write the id of the domain to the
[identity] default_domain_id parameter in keystone.conf, using the
keystone_domain { is_default => true } resource.

2) Allows user to specify a separate service domain for service accounts
One of the use cases for v3 is the ability to have the service account
keystone_user resources (e.g. glance, cinder, etc.) to be in a separate
domain from the regular user accounts.  keystone::roles::admin allows
to specify this domain.

Implements: blueprint api-v3-support

Change-Id: I3c69ee343e4335f4ce86cd72d6a9e664f7abcf25