Only declare openstackclient class if updating version
Declaring the openstackclient class from keystone
means that no other puppet class can include the client
libs.

This is not a good idea since the client libraries will
commonly be added to all roles.

This patch ensures that the class is only included and
not declared unless a value other than the default of
ensure=>present needs to be set for the package.

This case was already covered by the existing spec tests
so no further tests were added.

Change-Id: Ifda7c47b28a57aa8495d643c7dd81896258c0d18