support for keystone v3 api - keystone::resource::authtoken
This patch implements these parts of the blueprint:

1) Adds a new resource keystone::resource::authtoken - this manages
keystone authentication for other puppet modules by managing the
[keystone_authtoken] section in the config.  Encapsulating the
authentication configuration in this resource will ensure that we
will be able to support the several new types of authentication coming
in current and future Keystone releases.  This supports all of the
authentication options specified at:
http://www.jamielennox.net/blog/2015/02/17/loading-authentication-plugins/

The username and project_name can be passed in with domain, as
"username::domainname" or "project::domainname".  The resource will strip
off the domain part and use it for the user_domain_name or
project_domain_name if no other domain is specified.

Implements: blueprint api-v3-support

Change-Id: Ib37f11ae5949e6add10041c108a92fd3de553557