support for keystone v3 api - the user resource
Implements: blueprint api-v3-support

This patch implements these parts of the blueprint:
1) Implements the domain aware keystone_user resource.
Adds support for a :domain property of a user, and allows
the domain to be specified as "name::domain" in the :name property.
The new code tries to use the id as much as possible, since it is
the id that uniquely identifies the user among all of
the domains, and using id means that name + domain are not required.
The use of openstack project and role has changed somewhat because
the command line arguments and output are different in v3.
openstack user create --project $project does _not_ add the user
as a member of $project, so that must be done explicitly in
keystone_user.create in order not to break existing manifests that
expect that behavior.

Change-Id: Ie64e03705fa861d3ba5b91b79574a943841319a8