acceptance: run keystone in a WSGI server
Running keystone via eventlet is deprecated as of Kilo in favor of
running in a WSGI server (e.g. mod_wsgi).
Support for keystone under eventlet will be removed in the "M"-Release.

This patch test an upgrade between eventlet to wsgi.
In the future, the eventlet test will be dropped in favor of wsgi which
will be default.

Change-Id: Ieb07473725fbd51aac45e90c9aac1e819b944ee0