Resource keystone_identity_provider for Keystone.
Used for federation, this implement [1]

The remote-id parameter is missing from openstack client kilo release on
most distribution so this provider will work from liberty on.

The hardest part of the implementation is the support for the remote-id
and remote-id-file which are parameters which match the openstack cli.

First they are mutually exclusive.  Second, a file full of remote-id is
passed to the remote-id-file.  This file must be parsed to check if it's
in sync.  If the current format change, ie one id by line, it will need
to be changed.  Third the parsing of the shell output of the remote-id
is rather challenging before openstack cli 1.9.0[2].  This code handles
pre and post 1.9.0.

[1] http://specs.openstack.org/openstack/keystone-specs/api/v3/identity-api-v3-os-federation-ext.html#identity-provider-api
[2] https://bugs.launchpad.net/python-openstackclient/+bug/1478995

Change-Id: I4003d1a16d8ac3694f350c82b7b990f374a4a7b6