v3: make sure default domain is created before any other resource
When using Keystone v3 and domains, we need to make sure the default
domain (if its name if not 'Default') is created before any other
domain scoped resource.

By creating a new anchor, we can add the requirement in Keystone types
that need this dependency.
If the default domain name is not modified, the Anchor won't be in the
catalog but it's not an issue when using 'autorequire' in Puppet types.

This patch also change the default domains in acceptance tests, so we
can actually test the feature and make sure resources are created after
having the default domain created.

Change-Id: I2870eaa98f816c92df901ed2fa92e8db89b67656
Closes-bug: #1478037