WSGI: use real service name in restart_keystone Exec
Before, when running the Exec which restart Keystone service, we used
$service_name but this is wrong because service_name can be keystone,
openstack-keystone or httpd.
On Ubuntu/Debian, Apache service name is not httpd but apache2.

So the service could never be restarted if we were 1/ running
Ubuntu/Debian 2/ changing the default domain name.

This patch looks up in puppetlabs-apache the real Apache service name
when $service_name is httpd (if WSGI is enabled and eventlet disabled).

That way, we are sure the service name that will be used will the right
one.
Also add some unit tests that validate this change.

Change-Id: I85226d862b63e227867ed43bc35d7fefc9e424cd
Closes-bug: #1479783