if running eventlet, send deprecation warning
The Keystone development team decided to deprecate deployment of
Keystone under Eventlet during the Kilo cycle. Support for deploying
under eventlet will be dropped as of the M-release of OpenStack [1].

This patch aims to send a warning if running eventlet so users are aware
we will drop eventlet support right after stable/liberty release.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-February/057359.html

Change-Id: I826b8c7a312d8c41f15eb5b9f35bb6d46d414f4f