acceptance/eventlet: make sure apache is stopped
When testing upgrades, we have some cases where we apply the wsgi
manifest first and then eventlet. The problem is apache which is started
so keystone cannot be started because ports are already in use.

So this patch installs Apache on eventlet manifest and makes sure the
service is stopped before trying to start keystone with eventlet.

Change-Id: I8910c24e162924adf2b02485bd5cae24faba95ff