Support for multiple ldap backend.
This enable the user to inject multiple ldap backend configurations into
keystone.

Currently the ldap configuration is modeled through a class and injected
inside keystone.conf.  In a multiple domains environment, this prevents
the user to create a ldap configuration by domain.

A deprecation warning is added to the current ldap class.  This class is
not using the define as doing so would automatically trigger a restart
of the keystone server.  This would be unexpected by the openstack
operator and would certainly be seen as a bug.  This imply a lot of code
duplication but is required to make a smooth transition.

Change-Id: I75307d4a04510d8ba1a24663b1724849ea5b48f5