domain name from id lookups return empty
If the given domain id is not available in the @domain_hash,
look it up in Keystone and add it to the hash.  Do not assume
the @domain_hash can be populated once.  It must take into
account domains that are created later.
This consolidates the various bits of code that reset the class
variables in the provider superclass into spec_helper.rb - we must
be consistent about resetting the variables to ensure the spec
tests return consistent results.

Closes-Bug: #1492843
Change-Id: Ic475ca7f33c49784a0e5f4a432220cbf8b3ce635