able to use wrong resource name
In the provider create() methods, @property_hash is set to the
hash return of the request(type, 'create') command.  This sets
[:name] to the short name without the domain, which means the
provider.name method will not always return the resource name.
It must always return the resource name, so replace it with
the resource[:name].
Also, the user :tenant name value must be the one specified in
the resource definition.

Closes-Bug: #1492848
Change-Id: Ic8083b7bfc996cb73a3d128f43f4e36f96b69c36